wd: v1: fix the block memory problem
When we use the 'wd_bmm' to alloc and free memory, if the blk_size we
required is bigger than the blk_size in the pool, it still allocates
memory successfully.
So compare the blk_size what we required with the available size
in the pool first, then decide whether to allocate or not.

Change-Id: I7500725e6a7ea1e598ed5791564171ea84d97643
Signed-off-by: y00484669 <y00484669@huawei.com>
Signed-off-by: Meng Yu <y00484669@huawei.com>
Reviewed-on: http://arm-gerrit.turing-ci.hisilicon.com:8080/22586
Tested-by: public TuringEE <turingee@huawei.com>
Reviewed-by: xuzibo 00494022 <xuzaibo@huawei.com>
Reviewed-by: public TuringEE <turingee@huawei.com>
Signed-off-by: Haojian Zhuang <haojian.zhuang@linaro.org>